WEBAPP-157: Move ConditionalLink to common/components/
